Create conditional.java
This is the first submission of the java-dsa course. In these codes, a high type of Java programming will be used to know DSA better we need to go deep into the basics and main functions used in Java. in this code conditional statements like if-else,else-if, and else are used using scanner class.